add new table reserva_detalle for listing servicios

+++ php/reserva_eliminar.php
@@ -2,39 +2,70 @@
 require("main.php");
 $reserva_id_del = limpiar_cadena($_POST["eliminar"]);
 
-//verifica cliente
-$check_reserva = con();
-$check_reserva=$check_reserva->query("SELECT * FROM reserva WHERE reserva_id = '$reserva_id_del'");
 
-if($check_reserva->rowCount()==1){
-    $datos = $check_reserva->fetch();
+$check_detalle = con();
+$check_detalle=$check_detalle->query("SELECT * FROM reserva_detalle WHERE reserva_id = '$reserva_id_del'");
+    
+if($check_detalle->rowCount()>=1){
+    $detalles = $check_detalle->fetchAll();
 
-    $eliminar_reserva = con();
-    $eliminar_reserva=$eliminar_reserva->prepare("DELETE FROM reserva WHERE reserva_id=:id");
+    $eliminar_detalle = con();
+    $eliminar_detalle=$eliminar_detalle->prepare("DELETE FROM reserva_detalle WHERE reserva_id=:id");
     //filtro prepare para evitar inyecciones sql xss
 
-    $eliminar_reserva->execute([":id"=> $reserva_id_del]);
-    if($eliminar_reserva->rowCount()==1){
+    $eliminar_detalle->execute([":id"=> $reserva_id_del]);
+    if($eliminar_detalle->rowCount()>=1){
 
-        echo '
-        <div class="alert alert-success" role="alert">
-            <strong>Reserva eliminada!</strong><br>
-            La Reserva ha sido eliminada exitosamente.
-        </div>';
+        $detalles = true;
+
+    } else {
+
+        $detalles = false;
+    }
+    $eliminar_detalle=null;
+}
+
+if($detalles){
+    //verifica reserva
+    $check_reserva = con();
+    $check_reserva=$check_reserva->query("SELECT * FROM reserva WHERE reserva_id = '$reserva_id_del'");
+
+    if($check_reserva->rowCount()==1){
+        $datos = $check_reserva->fetch();
+
+        $eliminar_reserva = con();
+        $eliminar_reserva=$eliminar_reserva->prepare("DELETE FROM reserva WHERE reserva_id=:id");
+        //filtro prepare para evitar inyecciones sql xss
+
+        $eliminar_reserva->execute([":id"=> $reserva_id_del]);
+        if($eliminar_reserva->rowCount()==1){
+
+            echo '
+            <div class="alert alert-success" role="alert">
+                <strong>Reserva eliminada!</strong><br>
+                La Reserva ha sido eliminada exitosamente.
+            </div>';
+        } else {
+            echo '
+            <div class="alert alert-danger" role="alert">
+                <strong>¡Ocurrió un error inesperado!</strong><br>
+                No se pudo eliminar la Reserva, inténtelo nuevamente.
+            </div>';
+        }
+        $eliminar_reserva=null;
     } else {
         echo '
         <div class="alert alert-danger" role="alert">
             <strong>¡Ocurrió un error inesperado!</strong><br>
-            No se pudo eliminar la Reserva, inténtelo nuevamente.
+            La Reserva que intenta eliminar no existe.
         </div>';
     }
-    $eliminar_reserva=null;
+
+    $check_reserva=null;
 } else {
     echo '
     <div class="alert alert-danger" role="alert">
         <strong>¡Ocurrió un error inesperado!</strong><br>
-        La Reserva que intenta eliminar no existe.
+        No se pudo eliminar la Reserva, inténtelo nuevamente.
     </div>';
 }
-
-$check_reserva=null;
